update card_decks_controller actions(index,create,destroy), wrote rspec test

+++ app/controllers/api/v1/card_decks_controller.rb
@@ -2,23 +2,20 @@ class Api::V1::CardDecksController < ApplicationController
   
   # GET /api/v1/card_decks
   def index
-    card_decks = CardDeck.all
-    card_decks_json = ActiveModel::Serializer::CollectionSerializer.new(card_decks, each_serializer: CardSerializer)
+    authorize CardDeck
+    
+    card_decks = CardDeck.joins(:card, :deck).where(cards: { user_id: current_user.id }).where(decks: { user_id: current_user.id }).order(id: :desc)
+    card_decks_json = ActiveModel::Serializer::CollectionSerializer.new(card_decks, each_serializer: CardDeckSerializer)
     render json: { card_decks: card_decks_json }, status: :ok
   end
   
-  # POST /api/v1/decks/:deck_id/cards/:card_id/card_decks
+  # POST /api/v1/card_decks
   def create
-    deck = Deck.find(params[:deck_id])
-    deck.user = current_user
-    card = Card.find(params[:card_id])
-    card.user = current_user
+    card_deck = CardDeck.find_or_create_by(card_deck_params)
     
-    card_deck = CardDeck.new(card_deck_params)
-    card_deck.card_id = card.id
-    card_deck.deck_id = deck.id
+    authorize card_deck
     
-    if card_deck.save
+    if card_deck
       card_deck_json = CardDeckSerializer.new(card_deck).as_json
       render json: { card_deck: card_deck_json }, status: :created
     else
@@ -26,13 +23,13 @@ class Api::V1::CardDecksController < ApplicationController
     end
   end
   
-  # DELETE /api/v1/decks/:deck_id/cards/:card_id/card_decks/:id
+  # DELETE /api/v1/card_decks/:id
   def destroy
-    deck = Deck.find(params[:deck_id])
-    card = Card.find(params[:card_id])
     card_deck = CardDeck.find(params[:id])
-  
     card_deck_json = CardDeckSerializer.new(card_deck).as_json
+    
+    authorize card_deck
+    
     if card_deck.destroy
       render json: { card_deck: card_deck_json }, status: :ok
     else
